requirements.txt
# Jira and APIs
requests==2.31.0
anthropic>=0.40.0

# Data processing
pyyaml==6.0.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25

# Utilities
python-dateutil==2.8.2
rich==13.7.0

# Testing (development)
pytest==7.4.4
pytest-mock==3.12.0
